Updating containers: validate a 'put' request as if it is a 'patch'
That means that you can issue a 'put' and supply 'barcode' as a field
as long as you're specifying the same barcode it had before.
But it also means that you can issue a 'put' request that will leave
out information like a 'patch', but may be processed like a 'put'.
Ideally, we should disable 'put' completely and only support 'patch'
requests.